neural_networks_project/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
tensorflow==2.13.0
numpy==1.24.3
Pillow==10.0.1
redis==4.6.0
opencv-python==4.8.1.78
scikit-learn==1.3.0
matplotlib==3.7.2
